BUG FIX: bug related to casting between unsigned and signed has been fixed
BOSFrame method implicitly casts int to uint16_t while applying bitwise negation operator on CODE_ONLY_MASK. Now, we explicitly casts it to unsigned first and then cast it back to uint16_t. Some debug message has been fixed to display hex value in proper format by casting uint8_t to int,

+++ src/hexabitz/BOSFrame.cpp
@@ -17,7 +17,7 @@ void hstd::Frame::sanitize(void)
 
 void hstd::Frame::setCodeOnly(uint16_t newCode)
 {
-	const uint16_t clearMask = ~ CODE_ONLY_MASK;
+	const uint16_t clearMask = uint16_t(~ (unsigned(CODE_ONLY_MASK)));
 	code &= clearMask;
 	code |= (newCode & CODE_ONLY_MASK);
 }
